Implement feature-live module
