--- app/src/main/res/layout/item_live_message.xml
@@ -1,111 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-    Copyright 2022 Stream.IO, Inc. All Rights Reserved.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="data"
-            type="com.getstream.sdk.chat.adapter.MessageListItem.MessageItem" />
-
-        <variable
-            name="localDate"
-            type="String" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingTop="@dimen/margin_tiny"
-        android:paddingBottom="@dimen/margin_small">
-
-        <io.getstream.chat.android.ui.avatar.AvatarView
-            android:id="@+id/avatarView"
-            style="?attr/streamUiMessageListItemAvatarStyle"
-            android:layout_width="42dp"
-            android:layout_height="42dp"
-            android:layout_marginTop="@dimen/margin_tiny"
-            app:layout_constraintStart_toStartOf="@id/marginStart"
-            app:layout_constraintTop_toTopOf="parent"
-            app:userData="@{data.message.user}"
-            app:streamUiAvatarBorderWidth="0dp"
-            app:streamUiAvatarOnlineIndicatorColor="@color/stream_ui_accent_green"
-            app:streamUiAvatarOnlineIndicatorEnabled="true"
-            app:streamUiAvatarOnlineIndicatorPosition="topRight"
-            tools:src="@tools:sample/avatars" />
-
-        <TextView
-            android:id="@+id/usernameTextView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/margin_tiny"
-            android:text="@{data.message.user.name}"
-            android:textColor="@color/stream_ui_black"
-            android:textSize="16sp"
-            android:textStyle="bold"
-            app:layout_constraintStart_toEndOf="@id/avatarView"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="Ironman" />
-
-        <TextView
-            android:id="@+id/messageTimeTextView"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/margin_small"
-            android:maxLines="1"
-            android:text="@{localDate}"
-            android:textColor="@color/text_mid_emphasis"
-            android:textSize="12sp"
-            app:layout_constraintBottom_toBottomOf="@id/usernameTextView"
-            app:layout_constraintEnd_toStartOf="@id/marginEnd"
-            app:layout_constraintHorizontal_bias="0"
-            app:layout_constraintStart_toEndOf="@id/usernameTextView"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="2 mins" />
-
-        <TextView
-            android:id="@+id/messageTextView"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:text="@{data.message.text}"
-            android:textColor="@color/stream_ui_text_color_primary"
-            android:textColorLink="@color/colorAccent"
-            android:textSize="14sp"
-            app:layout_constraintEnd_toEndOf="@id/marginEnd"
-            app:layout_constraintStart_toStartOf="@id/usernameTextView"
-            app:layout_constraintTop_toBottomOf="@id/usernameTextView"
-            tools:text="You can't transmit the alarm without quantifying the haptic AGP card!" />
-
-        <androidx.constraintlayout.widget.Guideline
-            android:id="@+id/marginStart"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layout_constraintGuide_begin="8dp" />
-
-        <androidx.constraintlayout.widget.Guideline
-            android:id="@+id/marginEnd"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layout_constraintGuide_end="8dp" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</layout>
