+++ core-uicomponents/src/main/res/menu/home_menu.xml
@@ -15,18 +15,18 @@
     limitations under the License.
 -->
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/fragment_live"
-        android:icon="@drawable/ic_live"
-        android:title="@string/tab_live" />
+  <item
+    android:id="@+id/fragment_live"
+    android:icon="@drawable/ic_live"
+    android:title="@string/tab_live" />
 
-    <item
-        android:id="@+id/fragment_channel_list"
-        android:icon="@drawable/ic_bubble"
-        android:title="@string/tab_chat" />
+  <item
+    android:id="@+id/fragment_channel_list"
+    android:icon="@drawable/ic_bubble"
+    android:title="@string/tab_chat" />
 
-    <item
-        android:id="@+id/fragment_mentions"
-        android:icon="@drawable/ic_mention"
-        android:title="@string/tab_mention" />
+  <item
+    android:id="@+id/fragment_mentions"
+    android:icon="@drawable/ic_mention"
+    android:title="@string/tab_mention" />
 </menu>
