+++ feature-chat/src/main/kotlin/io/getstream/avengerschat/feature/chat/binding/ViewBinding.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2022 Stream.IO, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.avengerschat.feature.chat.binding
+
+import androidx.databinding.BindingAdapter
+import io.getstream.avengerschat.feature.chat.R
+import io.getstream.chat.android.client.models.User
+import io.getstream.chat.android.ui.channel.list.header.ChannelListHeaderView
+
+internal object ViewBinding {
+
+  @JvmStatic
+  @BindingAdapter("channelListHeader")
+  fun bindChannelListHeader(channelListHeaderView: ChannelListHeaderView, user: User?) {
+    if (user != null) {
+      channelListHeaderView.apply {
+        setUser(user)
+        showOnlineTitle()
+        setOnlineTitle(context.getString(R.string.app_name))
+      }
+    }
+  }
+}
