+++ feature-home/src/main/res/layout/activity_home.xml
@@ -15,58 +15,58 @@
     limitations under the License.
 -->
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+  xmlns:app="http://schemas.android.com/apk/res-auto"
+  xmlns:tools="http://schemas.android.com/tools">
 
-    <data>
+  <data>
 
-        <variable
-            name="vm"
-            type="io.getstream.avengerschat.feature.home.common.HomeViewModel" />
+    <variable
+      name="vm"
+      type="io.getstream.avengerschat.feature.home.common.HomeViewModel" />
 
-        <variable
-            name="fragmentManager"
-            type="androidx.fragment.app.FragmentManager" />
-    </data>
+    <variable
+      name="fragmentManager"
+      type="androidx.fragment.app.FragmentManager" />
+  </data>
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/background"
-        app:statusBarColor="@{vm.parsedColor}"
-        tools:context=".view.home.HomeActivity">
+  <androidx.constraintlayout.widget.ConstraintLayout
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/background"
+    app:statusBarColor="@{vm.parsedColor}"
+    tools:context=".view.home.HomeActivity">
 
-        <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/container"
-            android:name="androidx.navigation.fragment.NavHostFragment"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:background="@color/background"
-            app:defaultNavHost="true"
-            app:layout_constraintBottom_toTopOf="@id/bottom_navigation"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:navGraph="@navigation/nav_graph" />
+    <androidx.fragment.app.FragmentContainerView
+      android:id="@+id/container"
+      android:name="androidx.navigation.fragment.NavHostFragment"
+      android:layout_width="match_parent"
+      android:layout_height="0dp"
+      android:background="@color/background"
+      app:defaultNavHost="true"
+      app:layout_constraintBottom_toTopOf="@id/bottom_navigation"
+      app:layout_constraintLeft_toLeftOf="parent"
+      app:layout_constraintRight_toRightOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:navGraph="@navigation/nav_graph" />
 
-        <com.google.android.material.bottomnavigation.BottomNavigationView
-            android:id="@+id/bottom_navigation"
-            android:layout_width="0dp"
-            android:layout_height="@dimen/bottom_navigation_height"
-            android:layout_marginStart="0dp"
-            android:layout_marginEnd="0dp"
-            android:background="@{vm.parsedColor}"
-            app:isGone="@{!vm.visibleBottomNav}"
-            app:itemIconTint="@color/white"
-            app:itemTextColor="@color/white"
-            app:labelVisibilityMode="labeled"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:menu="@menu/home_menu"
-            app:navigation="@{fragmentManager}"
-            app:totalUnreadCount="@{vm.totalUnreadCount}" />
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+      android:id="@+id/bottom_navigation"
+      android:layout_width="0dp"
+      android:layout_height="@dimen/bottom_navigation_height"
+      android:layout_marginStart="0dp"
+      android:layout_marginEnd="0dp"
+      android:background="@{vm.parsedColor}"
+      app:isGone="@{!vm.visibleBottomNav}"
+      app:itemIconTint="@color/white"
+      app:itemTextColor="@color/white"
+      app:labelVisibilityMode="labeled"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toStartOf="parent"
+      app:menu="@menu/home_menu"
+      app:navigation="@{fragmentManager}"
+      app:totalUnreadCount="@{vm.totalUnreadCount}" />
 
-    </androidx.constraintlayout.widget.ConstraintLayout>
+  </androidx.constraintlayout.widget.ConstraintLayout>
 
 </layout>
