+++ feature-home/src/main/res/navigation/nav_graph.xml
@@ -16,61 +16,27 @@
 -->
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
   android:id="@+id/main_nav"
   app:startDestination="@id/fragment_live">
 
   <include app:graph="@navigation/dm_navigation" />
   <include app:graph="@navigation/chat_navigation" />
   <include app:graph="@navigation/user_navigation" />
+  <include app:graph="@navigation/live_navigation" />
 
-<!--  <fragment-->
-<!--    android:id="@+id/fragment_live"-->
-<!--    android:name="io.getstream.avengerschat.view.live.LiveFragment"-->
-<!--    tools:layout="@layout/fragment_live">-->
+  <!--  <fragment-->
+  <!--    android:id="@+id/fragment_mentions"-->
+  <!--    android:name="io.getstream.avengerschat.view.mention.MentionsFragment"-->
+  <!--    tools:layout="@layout/fragment_mentions">-->
 
-<!--    <action-->
-<!--      android:id="@+id/action_to_fragment_live_stream"-->
-<!--      app:destination="@id/fragment_live_stream"-->
-<!--      app:enterAnim="@anim/nav_enter_anim"-->
-<!--      app:exitAnim="@anim/nav_exit_anim"-->
-<!--      app:popEnterAnim="@anim/nav_enter_anim"-->
-<!--      app:popExitAnim="@anim/nav_exit_anim"-->
-<!--      app:popUpToInclusive="true" />-->
-
-<!--    <action-->
-<!--      android:id="@+id/action_to_fragment_user_profile_edit"-->
-<!--      app:destination="@id/fragment_user_profile_edit"-->
-<!--      app:enterAnim="@anim/nav_enter_anim"-->
-<!--      app:exitAnim="@anim/nav_exit_anim"-->
-<!--      app:popEnterAnim="@anim/nav_enter_anim"-->
-<!--      app:popExitAnim="@anim/nav_exit_anim"-->
-<!--      app:popUpToInclusive="true" />-->
-<!--  </fragment>-->
-
-<!--  <fragment-->
-<!--    android:id="@+id/fragment_live_stream"-->
-<!--    android:name="io.getstream.avengerschat.view.live.LiveStreamFragment"-->
-<!--    tools:layout="@layout/fragment_live_stream">-->
-
-<!--    <argument-->
-<!--      android:name="info"-->
-<!--      app:argType="io.getstream.avengerschat.core.model.LiveRoomInfo" />-->
-<!--  </fragment>-->
-
-<!--  <fragment-->
-<!--    android:id="@+id/fragment_mentions"-->
-<!--    android:name="io.getstream.avengerschat.view.mention.MentionsFragment"-->
-<!--    tools:layout="@layout/fragment_mentions">-->
-
-<!--    <action-->
-<!--      android:id="@+id/action_to_fragment_message_list"-->
-<!--      app:destination="@id/fragment_message_list"-->
-<!--      app:enterAnim="@anim/nav_enter_anim"-->
-<!--      app:exitAnim="@anim/nav_exit_anim"-->
-<!--      app:popEnterAnim="@anim/nav_enter_anim"-->
-<!--      app:popExitAnim="@anim/nav_exit_anim"-->
-<!--      app:popUpToInclusive="true" />-->
-<!--  </fragment>-->
+  <!--    <action-->
+  <!--      android:id="@+id/action_to_fragment_message_list"-->
+  <!--      app:destination="@id/fragment_message_list"-->
+  <!--      app:enterAnim="@anim/nav_enter_anim"-->
+  <!--      app:exitAnim="@anim/nav_exit_anim"-->
+  <!--      app:popEnterAnim="@anim/nav_enter_anim"-->
+  <!--      app:popExitAnim="@anim/nav_exit_anim"-->
+  <!--      app:popUpToInclusive="true" />-->
+  <!--  </fragment>-->
 
 </navigation>
