+++ feature-live/build.gradle.kts
@@ -1,3 +1,5 @@
+import io.getstream.Configuration
+
 @Suppress("DSL_SCOPE_VIOLATION")
 plugins {
   id(libs.plugins.android.library.get().pluginId)
@@ -8,11 +10,11 @@ plugins {
 }
 
 android {
-  compileSdk = io.getstream.Configuration.compileSdk
+  compileSdk = Configuration.compileSdk
 
   defaultConfig {
-    minSdk = io.getstream.Configuration.minSdk
-    targetSdk = io.getstream.Configuration.targetSdk
+    minSdk = Configuration.minSdk
+    targetSdk = Configuration.targetSdk
   }
 
   compileOptions {
@@ -31,19 +33,14 @@ dependencies {
   implementation(project(":core-uicomponents"))
 
   // feature modules
-  implementation(project(":feature-dm"))
-  implementation(project(":feature-chat"))
-  implementation(project(":feature-live"))
+  implementation(project(":feature-user"))
   implementation(project(":feature-home-common"))
 
   // data binding
   implementation(libs.bindables)
 
-  // custom views
-  implementation(libs.transformationLayout)
-
-  // bundler
-  implementation(libs.bundler)
+  // youtube player
+  implementation(libs.youtube.player)
 
   // hilt
   implementation(libs.hilt.android)
