+++ feature-live/src/main/AndroidManifest.xml
@@ -14,19 +14,4 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 -->
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/fragment_live"
-        android:icon="@drawable/ic_live"
-        android:title="@string/tab_live" />
-
-    <item
-        android:id="@+id/fragment_channel_list"
-        android:icon="@drawable/ic_bubble"
-        android:title="@string/tab_chat" />
-
-    <item
-        android:id="@+id/fragment_mentions"
-        android:icon="@drawable/ic_mention"
-        android:title="@string/tab_mention" />
-</menu>
+<manifest package="io.getstream.avengerschat.feature.live" />
