+++ feature-live/src/main/res/layout/fragment_live_stream.xml
@@ -1,0 +1,75 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+    Copyright 2022 Stream.IO, Inc. All Rights Reserved.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+-->
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto">
+
+  <data>
+
+    <variable
+      name="vm"
+      type="io.getstream.avengerschat.feature.home.common.HomeViewModel" />
+
+    <variable
+      name="video"
+      type="String" />
+  </data>
+
+  <androidx.constraintlayout.widget.ConstraintLayout
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/background">
+
+    <com.pierfrancescosoffritti.androidyoutubeplayer.core.player.views.YouTubePlayerView
+      android:id="@+id/youtubePlayerView"
+      android:layout_width="0dp"
+      android:layout_height="230dp"
+      android:clickable="false"
+      app:enableAutomaticInitialization="false"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:playYoutubeVideo="@{video}"
+      app:showFullScreenButton="false" />
+
+    <io.getstream.chat.android.ui.message.list.MessageListView
+      android:id="@+id/messageListView"
+      android:layout_width="0dp"
+      android:layout_height="0dp"
+      android:clipToPadding="false"
+      app:layout_constraintBottom_toTopOf="@+id/messageInputView"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintTop_toBottomOf="@id/youtubePlayerView"
+      app:streamUiBackgroundColor="@color/stream_ui_white_smoke"
+      app:streamUiFlagMessageConfirmationEnabled="true"
+      app:streamUiMessageTextColorDateSeparator="@color/white"
+      app:streamUiMessageTextSizeUserName="15sp"
+      app:streamUiMessageTextStyleUserName="bold"
+      app:streamUiMuteUserEnabled="false"
+      app:streamUiPinMessageEnabled="true" />
+
+    <io.getstream.chat.android.ui.message.input.MessageInputView
+      android:id="@+id/messageInputView"
+      android:layout_width="0dp"
+      android:layout_height="wrap_content"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toStartOf="parent"
+      app:streamUiAttachButtonEnabled="false" />
+
+  </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
