+++ feature-live/src/main/res/layout/item_stream_developers_channel.xml
