+++ feature-live/src/main/res/navigation/live_navigation.xml
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+    Copyright 2022 Stream.IO, Inc. All Rights Reserved.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+-->
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto"
+  xmlns:tools="http://schemas.android.com/tools"
+  android:id="@+id/dm_navigation"
+  app:startDestination="@+id/fragment_live">
+
+  <fragment
+    android:id="@+id/fragment_live"
+    android:name="io.getstream.avengerschat.feature.live.LiveFragment"
+    tools:layout="@layout/fragment_live">
+
+    <action
+      android:id="@+id/action_to_fragment_live_stream"
+      app:destination="@id/fragment_live_stream"
+      app:enterAnim="@anim/nav_enter_anim"
+      app:exitAnim="@anim/nav_exit_anim"
+      app:popEnterAnim="@anim/nav_enter_anim"
+      app:popExitAnim="@anim/nav_exit_anim"
+      app:popUpToInclusive="true" />
+
+    <action
+      android:id="@+id/action_to_fragment_user_profile_edit"
+      app:destination="@id/fragment_user_profile_edit"
+      app:enterAnim="@anim/nav_enter_anim"
+      app:exitAnim="@anim/nav_exit_anim"
+      app:popEnterAnim="@anim/nav_enter_anim"
+      app:popExitAnim="@anim/nav_exit_anim"
+      app:popUpToInclusive="true" />
+  </fragment>
+
+  <fragment
+    android:id="@+id/fragment_live_stream"
+    android:name="io.getstream.avengerschat.feature.live.LiveStreamFragment"
+    tools:layout="@layout/fragment_live_stream">
+
+    <argument
+      android:name="info"
+      app:argType="io.getstream.avengerschat.core.model.LiveRoomInfo" />
+  </fragment>
+
+</navigation>
