+++ settings.gradle.kts
@@ -31,3 +31,4 @@ include(":feature-chat")
 include(":feature-home")
 include(":feature-home-common")
 include(":feature-user")
+include(":feature-live")
